move custom fields to a dedicated folder
