+++ src/pages/Settings/content/formSections/Fields/CustomCheckbox.field.tsx
@@ -1,19 +1,16 @@
 import React, { Component } from 'react'
 import { InputField } from 'src/components/Form/Fields'
-import { Label, HiddenInput } from './elements'
+import { Label, HiddenInput } from '../elements'
 import { Image } from 'rebass'
 import Text from 'src/components/Text'
-import { ProfileTypeLabel } from 'src/models/user_pp.models'
 
 interface IProps {
   value: string
   index: number
-  onChange: (value: ProfileTypeLabel) => void
+  onChange: (index: number) => void
   isSelected: boolean
   imageSrc?: string
   btnLabel?: string
-  textLabel?: string
-  name: string
 }
 interface IState {
   showDeleteModal: boolean
@@ -22,7 +19,7 @@ interface IState {
 // validation - return undefined if no error (i.e. valid)
 const required = (value: any) => (value ? undefined : 'Required')
 
-class CustomRadioField extends Component<IProps, IState> {
+class CustomCheckbox extends Component<IProps, IState> {
   constructor(props: IProps) {
     super(props)
     this.state = {
@@ -31,26 +28,16 @@ class CustomRadioField extends Component<IProps, IState> {
   }
 
   render() {
-    const {
-      value,
-      imageSrc,
-      isSelected,
-      btnLabel,
-      textLabel,
-      name,
-    } = this.props
+    const { value, index, imageSrc, isSelected, btnLabel } = this.props
     return (
       <Label htmlFor={value} className={isSelected ? 'selected' : undefined}>
         <HiddenInput
           id={value}
-          name={name}
-          value={value}
-          type="radio"
+          name={value}
+          type="checkbox"
           component={InputField}
           checked={isSelected}
-          onChange={v => {
-            this.props.onChange(v.target.value)
-          }}
+          onChange={() => this.props.onChange(index)}
         />
         {imageSrc && <Image px={3} src={imageSrc} />}
         {btnLabel && (
@@ -58,14 +45,9 @@ class CustomRadioField extends Component<IProps, IState> {
             {btnLabel}
           </Text>
         )}
-        {textLabel && (
-          <Text px={1} my={1} txtcenter small>
-            {textLabel}
-          </Text>
-        )}
       </Label>
     )
   }
 }
 
-export { CustomRadioField }
+export { CustomCheckbox }
