+++ src/pages/Settings/content/formSections/UserInfos.section.tsx
@@ -12,7 +12,7 @@ import 'react-flags-select/scss/react-flags-select.scss'
 import styled from 'styled-components'
 import theme from 'src/themes/styled.theme'
 import { FieldArray } from 'react-final-form-arrays'
-import { Link } from './Link.field'
+import { Link } from './Fields/Link.field'
 import { ImageInputField } from 'src/components/Form/ImageInput.field'
 import { FlexSectionContainer } from './elements'
